[REF] Create single header
     - Chreated single header file to implement spine in raylib.
