+++ lib/include/spine/extension.h
@@ -321,10 +321,4 @@ int _spCurveTimeline_binarySearch (float *values, int valuesLength, float target
 }
 #endif
 
-// Extension functions needed to integrate raylib
-void drawSkeleton(spSkeleton* skeleton, Vector3 position);
-Texture2D* texture_2d_create(char *path);
-void texture_2d_destroy();
-
-
 #endif /* SPINE_EXTENSION_H_ */
