+++ lib/src/spine/Animation.c
