+++ lib/src/spine/AnimationState.c
