+++ lib/src/spine/AnimationStateData.c
