+++ lib/src/spine/Atlas.c
