+++ lib/src/spine/AtlasAttachmentLoader.c
