+++ lib/src/spine/Attachment.c
