+++ lib/src/spine/AttachmentLoader.c
