+++ lib/src/spine/Bone.c
