+++ lib/src/spine/BoundingBoxAttachment.c
