+++ lib/src/spine/ClippingAttachment.c
