+++ lib/src/spine/PathAttachment.c
