+++ lib/src/spine/SkeletonBinary.c
