+++ lib/src/spine/SkeletonClipping.c
