+++ lib/src/spine/SkeletonData.c
