+++ lib/src/spine/Slot.c
