+++ lib/src/spine/SlotData.c
