+++ lib/src/spine/Triangulator.c
