+++ lib/src/spine/VertexAttachment.c
