+++ lib/src/spine/VertexEffect.c
